Add delayBeforeResponding to defaultConfig.
This property allows a per-app response delay (the artificial delay before
returning the generated response).

Also adds a tick() helper that uses a promise that resolves
on the next iteration of the event loop and artifically
advancing (via jasmine.clock()) a given number of ms.
Useful for dealing with promises that jasmine doesn't manage
to make synchronous despite an installed jasmine.clock().