Add blog example.
A more realistic example than that in index.html. Shows a simple blog that
relies on a server for: canPost, posts, title.

The mockBackend has a couple of GET endpoints, and a simple POST endpoint that
errors if a post contains Poop.

To avoid cluttering the example HTML, load.js is used to load pre-req's like React and Babel
before using Babel to execute the app.